feat: execute deployments in apiserver
