+++ src/apiserver/api_server_impl.h
@@ -61,6 +61,9 @@ class APIServerImpl : public APIServer {
     void RegisterGetDB();
     void RegisterGetTable();
 
+    void ExecuteProcedure(const InterfaceProvider::Params& param,
+            const butil::IOBuf& req_body, JsonWriter& writer);
+
     static bool Json2SQLRequestRow(const butil::rapidjson::Value& non_common_cols_v,
                                    const butil::rapidjson::Value& common_cols_v,
                                    std::shared_ptr<openmldb::sdk::SQLRequestRow> row);
